% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{mid_root_iqtree}
\alias{mid_root_iqtree}
\title{Mid-root IQ-TREE Function}
\usage{
mid_root_iqtree(tree_string, outgroup_label)
}
\arguments{
\item{tree_string}{The input IQ-TREE formatted tree string.}

\item{outgroup_label}{The label of the outgroup to extract.}
}
\value{
A modified IQ-TREE formatted tree string.
}
\description{
This function extracts a specified outgroup from an IQ-TREE formatted tree string and performs required operations.
}
